Use clustering to make path optimiser more efficient
Groups of lines where both endpoints are close together are clustered. Within a cluster, the lines are sorted in order of increasing y-intercept, which should work excellently for infill and skin lines. The clusters are then optimised with TSP. Not properly tested yet since elsewhere in the code it has a segfault, but I wanted to make these separate commits.

Contributes to issue CURA-329.