Cache edge length for faster normalising
The normal() function calculates the length of the vector that it's normalising. Since we normalise the same vector twice, we can do the expensive vSize() call once and re-use it.
Normally this would count as a micro-optimisation, but since I'm going to need this size anyway for my next commit I'm doing it anyway.

Contributes to issue CURA-3903.