Plan 3D paths
This is a change that has huge implications: Our layer plans now contain 50% more data, because their coordinates also have a Z. This requires us to convert the points to 2D every time we obtain points from the paths in the layer plans (which is what most of the changes in this commit entail) and convert it back to 3D when we add something to the path. In short, this will likely make the engine quite a lot slower.
On the flip side, we can now make 3D paths, and this allows for diagonal sanding. It should also make spiralize better, because we don't need to make that many exceptions any more for spiralize. However I haven't transformed the spiralize code to actually use 3D paths yet. I've left it 2D (converted the path vertices to 2D) and added the Z coordinate upon exporting to g-code, as it was.

This code is probably not optimal. There's quite a lot of places where it could make more sense to do calculations on 3D coordinates rather than 2D. In particular, we might send the layer view data to the front-end in 3D, and some of the distance checks we do right now could also be done in 3D. I've done it the easy way for now, which ensures that the behaviour should stay exactly the same by this commit. There are no g-code changes here.

Contributes to issue CURA-3903.