Fix differences between layers when sanding diagonally
Unfortunately this means that we have to generate new polygons and store them for just the diagonal sanding. Also, we currently calculate every offset twice. This is something that could be prevented by some form of second pass but not in our current structure. It is required for computing the difference with the offsetted polygon of the next layer, but the next layer may not have been calculated yet due to multithreading.

Contributes to issue CURA-3903.